# projects/neural-sync/artifacts/api-contracts.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NEURAL SYNC API",
    "version": "1.0.0",
    "description": "Multi-dimensional AI-driven developer–project compatibility engine. All protected routes require JWT Bearer authentication per ADR-002. The 500 ms p95 latency SLA applies to the synchronous POST /matches response per ADR-003. LLM explanations are generated asynchronously (Phase-1: Google Gemini; Claude API is a supported alternate per ADR-004); the initial response always contains a deterministic stub (≥50 chars) per ADR-003/ADR-004. Error envelope on all 4xx/5xx: {error_code, message, request_id}.",
    "contact": {
      "name": "NEURAL SYNC Engineering",
      "email": "engineering@neural-sync.example.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1",
      "description": "Local development server"
    },
    {
      "url": "https://neural-sync.example.com/api/v1",
      "description": "Production server"
    }
  ],
  "security": [
    { "BearerAuth": [] }
  ],
...
  "paths": {
...
    "/developers/{developer_id}/matches": {
      "get": {
        "tags": ["developers"],
        "summary": "Get ranked project recommendations for a developer",
        "description": "Returns top-K MatchRecord objects ranked by match_score descending. Used by the Developer Dashboard to render recommended project cards with explanation and risk tags. Developers may only query their own matches.",
        "operationId": "getDeveloperMatches",
        "parameters": [
          { "$ref": "#/components/parameters/DeveloperIdPath" },
          {
            "name": "top_k",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 10 },
            "description": "Maximum number of recommendations to return"
          },
          {
            "name": "min_score",
            "in": "query",
            "required": false,
            "schema": { "type": "number", "minimum": 0.0, "maximum": 1.0 },
            "description": "Filter results to match_score ≥ min_score"
          }
        ],
        "responses": {
          "200": {
            "description": "Ranked match recommendations",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DeveloperMatchesResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "403": {
            "description": "Forbidden",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "404": {
            "description": "Developer not found",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
...
    "/matches": {
      "post": {
        "tags": ["matches"],
        "summary": "Compute a five-dimension developer–project match score",
        "description": "Accepts inline DeveloperProfile and ProjectProfile objects. Computes MATCH_SCORE = w1·skill_score + w2·workstyle_score + w3·motivation_score + w4·timezone_score + w5·growth_score using the current WeightConfig (loaded fresh from DB per architecture decision). Returns within 500ms p95 per ADR-003. The explanation field is always a non-empty string ≥50 chars: a cached full LLM explanation on cache hit, or a deterministic algorithmic stub on cache miss. A background task enqueues the async LLM call (ADR-004: configurable provider; Phase-1 default is Google Gemini gemini-1.5-flash; Claude API claude-3-5-haiku is a supported alternate). The MatchRecord is persisted; use GET /matches/{match_id} to poll for the updated explanation.",
        "operationId": "createMatch",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MatchRequest" },
              "example": {
                "developer_profile": {
                  "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                  "skills": ["Python", "FastAPI", "PostgreSQL", "Docker"],
                  "experience_years": 5,
                  "preferred_stack": ["Python", "FastAPI"],
                  "work_style": [0.8, 0.6, 0.7, 0.9, 0.7, 0.8, 0.5, 0.4],
                  "motivation_vector": [0.9, 0.8, 0.5, 0.6, 0.7, 0.6, 0.8, 0.9],
                  "timezone": "Europe/Warsaw",
                  "availability_hours": 40,
                  "career_goals": ["technical leadership", "distributed systems"],
                  "project_history": [],
                  "is_self_reported": true
                },
                "project_profile": {
                  "id": "7e2b4f81-1234-4abc-9def-0a1b2c3d4e5f",
                  "name": "AI Platform Core",
                  "required_skills": ["Python", "FastAPI", "PostgreSQL"],
                  "team_structure": "Cross-functional agile squad of 6, async-first",
                  "workload_intensity": 0.7,
                  "innovation_level": 0.9,
                  "timezone_overlap_required": "UTC+1 to UTC+3",
                  "duration_weeks": 24,
                  "growth_opportunities": ["distributed systems", "ML pipeline", "technical leadership"]
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Match computed and persisted within 500ms p95 SLA",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MatchResponse" },
                "example": {
                  "match_id": "9b1deb4d-3b7d-4bad-9bdd-2b0d7b3dcb6d",
                  "developer_id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                  "project_id": "7e2b4f81-1234-4abc-9def-0a1b2c3d4e5f",
                  "match_score": 0.83,
                  "explanation": "Skill alignment: Strong overlap in Python, FastAPI and PostgreSQL covering 3 of 3 required skills. Behavioral fit: High collaboration and innovation scores align with the team's async-first, innovation-focused culture. Growth potential: Distributed systems and technical leadership goals match project's core growth opportunities.",
                  "explanation_source": "stub_pending",
                  "risks": ["Moderate timezone delta (UTC+2 vs UTC+1 to UTC+3) may compress synchronous windows"],
                  "growth_potential": ["Distributed systems architecture", "Technical leadership track", "ML pipeline exposure"],
                  "component_scores": {
                    "skill_score": 0.92,
                    "workstyle_score": 0.81,
                    "motivation_score": 0.87,
                    "timezone_score": 0.78,
                    "growth_score": 0.79
                  },
                  "weights_snapshot": { "w1": 0.30, "w2": 0.25, "w3": 0.20, "w4": 0.10, "w5": 0.15 },
                  "vector_search_degraded": false,
                  "behavioral_data_unavailable": false,
                  "created_at": "2026-06-26T10:00:00Z"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "401": {
            "description": "Unauthorized",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "422": {
            "description": "Validation error (e.g. work_style vector wrong length)",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "503": {
            "description": "Service temporarily unavailable — DB connection pool exhausted",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying",
                "schema": { "type": "integer", "example": 2 }
              }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "error_code": "SERVICE_UNAVAILABLE",
                  "message": "Connection pool exhausted. Please retry after 2 seconds.",
                  "request_id": "req_xyz001"
                }
              }
            }
          }
        }
      }
    },
...
    "/matches/rescore": {
      "post": {
        "tags": ["matches"],
        "summary": "Trigger batch match re-scoring",
        "description": "Re-computes MATCH_SCORE for all or a subset of matches using the current WeightConfig. Enables on-demand re-optimization after weight changes. Returns HTTP 202 Accepted; processing is asynchronous. Manager role required.",
        "operationId": "rescoreMatches",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RescoreRequest" }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Re-score job accepted and queued",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AsyncJobResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "403": {
            "description": "Forbidden — manager role required",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
    "/matches/{match_id}": {
      "get": {
        "tags": ["matches"],
        "summary": "Get a match record by ID",
        "description": "Returns the full MatchRecord. If explanation_source='stub_pending', poll this endpoint or GET /matches/{match_id}/explanation to check for the completed LLM explanation. Developers may only retrieve their own match records.",
        "operationId": "getMatch",
        "parameters": [
          { "$ref": "#/components/parameters/MatchIdPath" }
        ],
        "responses": {
          "200": {
            "description": "Match record found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MatchRecord" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "403": {
            "description": "Forbidden",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "404": {
            "description": "Match not found",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
    "/matches/{match_id}/explanation": {
      "get": {
        "tags": ["matches"],
        "summary": "Poll for the LLM-generated explanation",
        "description": "Returns the current explanation text and its source. Frontend shows a loading indicator while explanation_source='stub_pending' and polls this endpoint. Once explanation_source transitions to 'claude_async', 'claude_cached', or 'stub_permanent', polling may stop. Never exposes raw work_style or motivation_vector values.",
        "operationId": "getMatchExplanation",
        "parameters": [
          { "$ref": "#/components/parameters/MatchIdPath" }
        ],
        "responses": {
          "200": {
            "description": "Explanation returned (may still be a stub if the LLM call is pending)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ExplanationResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "403": {
            "description": "Forbidden",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "404": {
            "description": "Match not found",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token from POST /auth/login. Algorithm: HS256. Payload: {sub: user_id, role: 'developer'|'manager', iat, exp}. TTL: 1h. Secret loaded from env NEURAL_SYNC_JWT_SECRET. Per ADR-002: store in memory (not localStorage) on the React frontend; refresh token stored in HttpOnly cookie."
      }
    },
    "parameters": {
      "DeveloperIdPath": {
        "name": "developer_id",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" },
        "description": "Developer unique identifier (UUID v4)"
      },
...
      "MatchIdPath": {
        "name": "match_id",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" },
        "description": "Match record unique identifier (UUID v4)"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "description": "Standard error envelope returned on all 4xx and 5xx responses.",
        "required": ["error_code", "message", "request_id"],
        "properties": {
          "error_code": {
            "type": "string",
            "description": "Machine-readable error code. Examples: NOT_FOUND, INVALID_WEIGHTS, VALIDATION_ERROR, ERASURE_FAILED, SERVICE_UNAVAILABLE, FORBIDDEN, UNAUTHORIZED, CONFLICT"
          },
          "message": {
            "type": "string",
            "description": "Human-readable description of the error, safe to display in developer tooling"
          },
          "request_id": {
            "type": "string",
            "description": "Unique request trace identifier for correlating logs"
          }
        }
      },
...
      "DeveloperMatchesResponse": {
        "type": "object",
        "required": ["developer_id", "matches", "total"],
        "properties": {
          "developer_id": { "type": "string", "format": "uuid" },
          "matches": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/MatchRecord" },
            "description": "Ranked match records, highest match_score first"
          },
          "total": {
            "type": "integer",
            "minimum": 0,
            "description": "Total number of match records for this developer (before top_k filtering)"
          }
        }
      },
...
      "RescoreRequest": {
        "type": "object",
        "description": "Optional body for POST /matches/rescore. Omit entirely to re-score all developers.",
        "properties": {
          "developer_ids": {
            "type": ["array", "null"],
            "items": { "type": "string", "format": "uuid" },
            "description": "Scope re-scoring to these developer IDs. null or omitted = re-score all developers."
          }
        }
      },
...
    }
  }
}